fix: remove '(UTC)' from date strings rendered on server
The '(UTC)' was appended to the formatted dates present in the pages
rendered in the server and sent to clients. UTC dates are assumed in the
server, but they are usually not compatible with clients timezones.
Hence, on client side, those strings were substituted with the correct
local date strings.

The indication of '(UTC)', although indicated that the date were not
considerng the timezone, produced a flick on the client. So it has beeh
removed.